Start Bounce Pad (Not Working Check Comments)
Doesn't always activate the bounce pad when you enter it. After testing my current theory is that the y velocity is being reset somewhere in the code. I suspect that as the trigger is called asynchronously to the rest of the code if it is called between the Move function and this reset it doesn't activate properly. Current suspect is line 254 in the PLayer Controller.