Some boss fixes (Check comments)
changes where made into the boss testing scene as the other scene wasn't working correctly

made the boss ram the player can adjust the timing of it to make it better.

made the boss fall down and stop bouncing the player when beaten.

wasn't able to get cracks working yet. tried using a png as a material but got left with odd gaps where it wrapped or tiled

need to add gaurds back into the arena to make it harder to just cheese